Refactor CommandGroups into separate files
