+++ src/main/kotlin/frc/robot/RobotContainer.kt
@@ -6,10 +6,12 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser
 import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.Commands
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController
+import frc.robot.commandGroups.IntakeCommands
 import frc.robot.scoreState.AmpState
 import frc.robot.scoreState.ClimbState
 import frc.robot.scoreState.ScoreState
 import frc.robot.scoreState.ShootState
+import frc.robot.subsystems.intake.Intake
 import frc.robot.subsystems.swerve.SwerveDrive
 
 /**
@@ -52,8 +54,8 @@ object RobotContainer {
         ControllerInputs.operatorController().x().onTrue(Commands.runOnce({ currentState = ampState }))
         ControllerInputs.operatorController().y().onTrue(Commands.runOnce({ currentState = climbState }))
 
-        ControllerInputs.driverController().rightTrigger().whileTrue(CommandGroups.intake())
-        ControllerInputs.driverController().rightBumper().whileTrue(CommandGroups.outtake())
+        ControllerInputs.driverController().rightTrigger().whileTrue(IntakeCommands.intake())
+        ControllerInputs.driverController().rightBumper().whileTrue(IntakeCommands.outtake())
 
         ControllerInputs.operatorController().x().whileTrue(Intake.getInstance().reset())
