+++ src/main/kotlin/frc/robot/commandGroups/ShootOverStageConstants.kt
@@ -1,14 +1,14 @@
 package frc.robot.commandGroups
 
-import com.pathplanner.lib.util.GeometryUtil
-import edu.wpi.first.math.geometry.Translation2d
 import edu.wpi.first.units.Angle
 import edu.wpi.first.units.Measure
 import edu.wpi.first.units.Units
 import edu.wpi.first.units.Velocity
+import edu.wpi.first.math.geometry.Translation2d
+import com.pathplanner.lib.util.GeometryUtil
 import frc.robot.Constants
 
-object CommandGroupsConstants {
+object ShootOverStageConstants {
     val HOOD_ANGLE_SUPER_POOP: Measure<Angle> = Units.Degrees.of(110.0)
     val SHOOTER_VELOCITY_SUPER_POOP: Measure<Velocity<Angle>> = Units.RotationsPerSecond.of(77.07)
     val CONVEYOR_VELOCITY_SUPER_POOP: Measure<Velocity<Angle>> = Units.RotationsPerSecond.of(60.0)
