+++ src/main/kotlin/frc/robot/subsystems/climb/Climb.kt
@@ -5,7 +5,6 @@ import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.Commands
 import edu.wpi.first.wpilibj2.command.StartEndCommand
 import edu.wpi.first.wpilibj2.command.SubsystemBase
-import frc.robot.commandGroups.CommandGroups
 import org.littletonrobotics.junction.AutoLogOutput
 import org.littletonrobotics.junction.Logger
 import java.util.function.DoubleSupplier
